+++ example_test.go
@@ -1,16 +1,17 @@
 package logrotate
 
-import (
-	"log"
-)
+import "fmt"
 
-// 使用标准库 log 时，在应用启动时把 Logger 传给 SetOutput 即可。
+// New 可以用 Functional Options 创建 Logger。
 func Example() {
-	log.SetOutput(&Logger{
-		Filename:   "/var/log/myapp/foo.log",
-		MaxSize:    500, // MB
-		MaxBackups: 3,
-		MaxAge:     28,   // 天
-		Compress:   true, // 默认关闭
-	})
+	l := New(
+		WithFilename("/var/log/myapp/foo.log"),
+		WithMaxSize(500),
+		WithMaxBackups(3),
+		WithMaxAge(28),
+		WithCompress(true),
+	)
+
+	fmt.Println(l.Filename, l.MaxSize, l.MaxBackups, l.MaxAge, l.Compress)
+	// Output: /var/log/myapp/foo.log 500 3 28 true
 }